Reduce size of the packaged helm chart
The `docs` and the `examples` folder consume over 700 kbytes. A packaged
helm chart is still 240kbyte. This is way too big. The chart does not
need to contain that much bloat when used to deploy software.

By having the two folders in the .helmignore file, building the helm
chart leads to a 21kbyte file.