This commit changes the schema of the target file names in various aspects:
* All .pcap files are put into a subfolder grouped by day. This makes
  deleting old data from any remote target much easier. It also reduces
  stress on any file explorer which do not have to spend CPU cycles to
  render 20k files.

* All .pcap files are compressed by .lz4 before uploaded. Usually, this
  reduces the file size to 1/3 of the original size.

* The date part of the timestamp changes from "YYYY_MM_DD" to
  "YYYY-MM-DD".

Underneath, the way how parts of the filename is deduced is changed as
well. The timestamp is based upon the "change date" of the file. stat(1)
is able to print the proper timestamp. This makes any sed-orgy
irrelevant while at the same time opens the door to define the folder
name. That timestamp is then used to construct a properly filename based
upon date parts. This makes any sed-orgy to construct something similar
obsolete.