Fix filename generation in rclone script
The `target` variable is not initialized / reset to "" on each iteration
of the loop. Thus, $target kept grewing under some circumstances
(target=$target).

This commit addresses the issue. Ontop, it moves the actual file
handling code out of the "watch for new files" loop. Also, it surrounds
the detected filenames with quotes in order to handle filenames with
spaces properly (not expected, but still).